# clients/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "CreateGroupCommand": {
        "properties": {
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "isDefault": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "roleIds": {
            "items": {
              "type": "string"
            },
            "type": [
              "null",
              "array"
            ]
          }
        },
        "required": [
          "name",
          "description",
          "isDefault",
          "roleIds"
        ],
        "type": "object"
      },
...
      "GroupDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "isSystemGroup": {
            "type": "boolean"
          },
          "memberCount": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "roleIds": {
            "items": {
              "type": "string"
            },
            "type": [
              "null",
              "array"
            ]
          },
          "roleNames": {
            "items": {
              "type": "string"
            },
            "type": [
              "null",
              "array"
            ]
          }
        },
        "type": "object"
      },
...
      "PagedResponseOfGroupDto": {
        "properties": {
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GroupDto"
            },
            "type": "array"
          },
          "pageNumber": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "totalCount": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "Json Web Token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "contact@example.com",
      "name": "Boilerplate",
      "url": "https://example.com/contact"
    },
    "description": "Modular, multitenant SaaS API.",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Boilerplate API",
    "version": "v1"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/identity/groups": {
      "get": {
        "description": "Retrieve the current tenant's groups, ordered by name. Pageable via PageNumber/PageSize (at most 100); filterable via Search (case-insensitive substring against name and description).",
        "operationId": "ListGroups",
        "parameters": [
          {
            "in": "query",
            "name": "PageNumber",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "PageSize",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "Sort",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "Search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfGroupDto"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List groups (paged)",
        "tags": [
          "Identity"
        ]
      },
      "post": {
        "description": "Create a new group with optional role assignments.",
        "operationId": "CreateGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDto"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create a new group",
        "tags": [
          "Identity"
        ]
      }
    },
...
  },
...
}